removed StoneId, refactored namespaces, restructured files
