+++ include/digital_curling/detail/constants.hpp
@@ -1,0 +1,25 @@
+﻿#ifndef DIGITAL_CURLING_CONSTANTS_HPP
+#define DIGITAL_CURLING_CONSTANTS_HPP
+
+#include <cstdint>
+
+namespace digital_curling {
+
+
+
+/// <summary>
+/// 円周率
+/// </summary>
+constexpr float kPi = 3.14159265359f;
+
+
+/// <summary>
+/// シミュレータ上に配置可能なストーンの最大数
+/// </summary>
+constexpr std::uint8_t kStoneMax = 16;
+
+
+
+} // namespace digital_curling
+
+#endif // DIGITAL_CURLING_CONSTANTS_HPP
