+++ include/digital_curling/detail/game/moves/concede.hpp
@@ -1,0 +1,19 @@
+﻿#ifndef DIGITAL_CURLING_GAME_MOVES_CONCEDE_HPP
+#define DIGITAL_CURLING_GAME_MOVES_CONCEDE_HPP
+
+namespace digital_curling::game::moves {
+
+
+
+/// <summary>
+/// 行動：コンシード
+/// </summary>
+struct Concede {
+    static constexpr char kType[] = "concede";
+};
+
+
+
+} // namespace digital_curling::game::moves
+
+#endif // DIGITAL_CURLING_GAME_MOVES_CONCEDE_HPP
