+++ include/digital_curling/detail/vector2.hpp
@@ -1,32 +1,10 @@
-﻿#ifndef DIGITAL_CURLING_COMMON_HPP
-#define DIGITAL_CURLING_COMMON_HPP
+﻿#ifndef DIGITAL_CURLING_VECTOR2_HPP
+#define DIGITAL_CURLING_VECTOR2_HPP
 
-#include <cstdint>
 #include <cmath>
-#include <array>
-#include <optional>
-#include "nlohmann/json.hpp"
+#include "json.hpp"
 
 namespace digital_curling {
-
-/// <summary>
-/// 円周率
-/// </summary>
-constexpr float kPi = 3.14159265359f;
-
-
-
-/// <summary>
-/// ストーンのインデックスに用いる型
-/// </summary>
-using StoneId = std::uint8_t;
-
-
-
-/// <summary>
-/// シミュレータ上に配置可能なストーンの最大数
-/// </summary>
-constexpr StoneId kStoneMax = 16;
 
 
 
@@ -105,6 +83,8 @@ constexpr Vector2 operator / (Vector2 v, float f)
     return { v.x / f, v.y / f };
 }
 
+
+
 } // namespace digital_curling
 
 
@@ -118,30 +98,5 @@ void from_json(nlohmann::json const&, Vector2 &);
 
 } // namespace digital_curling
 
-namespace nlohmann {
 
-// For serializing/deserializing std::optional to/from json
-template <typename T>
-struct adl_serializer<std::optional<T>> {
-    static void to_json(json & j, std::optional<T> const& opt)
-    {
-        if (opt) {
-            j = *opt;
-        } else {
-            j = nullptr;
-        }
-    }
-
-    static void from_json(json const& j, std::optional<T> & opt)
-    {
-        if (j.is_null()) {
-            opt = std::nullopt;
-        } else {
-            opt = j.get<T>();
-        }
-    }
-};
-
-} // namespace nlohmann
-
-#endif // DIGITAL_CURLING_COMMON_HPP
+#endif // DIGITAL_CURLING_VECTOR2_HPP
