+++ include/digital_curling/digital_curling.hpp
@@ -1,13 +1,27 @@
 ﻿#ifndef DIGITAL_CURLING_DIGITAL_CURLING_HPP
 #define DIGITAL_CURLING_DIGITAL_CURLING_HPP
 
-#include "detail/common.hpp"
+#include "detail/constants.hpp"
 #include "detail/coordinate.hpp"
+#include "detail/json.hpp"
+#include "detail/vector2.hpp"
 #include "detail/version.hpp"
-#include "detail/game/common.hpp"
-#include "detail/game/normal_dist_shot_randomizer.hpp"
-#include "detail/game/normal.hpp"
+#include "detail/game/apply_move.hpp"
+#include "detail/game/constants.hpp"
+#include "detail/game/result.hpp"
+#include "detail/game/setting.hpp"
+#include "detail/game/state.hpp"
+#include "detail/game/team.hpp"
+#include "detail/game/moves/concede.hpp"
+#include "detail/game/moves/shot.hpp"
+#include "detail/game/moves/time_limit.hpp"
+#include "detail/game/random/normal_dist_shot_randomizer.hpp"
+#include "detail/game/random/random_seed_normal_dist_shot_randomizer.hpp"
+#include "detail/game/random/shot_randomizer.hpp"
+#include "detail/simulation/simulator_fcv1_setting.hpp"
+#include "detail/simulation/simulator_setting.hpp"
 #include "detail/simulation/simulator.hpp"
-#include "detail/simulation/fcv1.hpp"
+#include "detail/simulation/stone_collision.hpp"
+#include "detail/simulation/stone_data.hpp"
 
-#endif
+#endif // DIGITAL_CURLING_DIGITAL_CURLING_HPP
